# Switch to ruff for faster static checks