--- a/dev/breeze/src/airflow_breeze/pre_commit_ids.py
+++ b/dev/breeze/src/airflow_breeze/pre_commit_ids.py
@@ -20,6 +20,7 @@
 #
 # IF YOU WANT TO MODIFY IT, YOU SHOULD MODIFY THE TEMPLATE
 # `pre_commit_ids_TEMPLATE.py.jinja2` IN the `dev/breeze/src/airflow_breeze` DIRECTORY
+from __future__ import annotations
 
 PRE_COMMIT_LIST = [
     "all",
@@ -79,7 +80,6 @@ PRE_COMMIT_LIST = [
     "flynt",
     "identity",
     "insert-license",
-    "isort",
     "lint-chart-schema",
     "lint-css",
     "lint-dockerfile",
@@ -89,15 +89,12 @@ PRE_COMMIT_LIST = [
     "lint-openapi",
     "mixed-line-ending",
     "pretty-format-json",
-    "pydocstyle",
     "python-no-log-warn",
-    "pyupgrade",
     "replace-bad-characters",
     "rst-backticks",
-    "run-flake8",
+    "ruff",
     "run-mypy",
     "run-shellcheck",
-    "static-check-autoflake",
     "trailing-whitespace",
     "ts-compile-and-lint-javascript",
     "update-black-version",
@@ -116,5 +113,4 @@ PRE_COMMIT_LIST = [
     "update-vendored-in-k8s-json-schema",
     "update-version",
     "yamllint",
-    "yesqa",
 ]