--- a/docs/spelling_wordlist.txt
+++ b/docs/spelling_wordlist.txt
@@ -754,6 +754,7 @@ IRSA
 isfile
 ish
 isn
+isort
 iterable
 iterables
 iteratively