--- a/images/breeze/output-commands-hash.txt
+++ b/images/breeze/output-commands-hash.txt
@@ -51,7 +51,7 @@ setup:version:123b462a421884dc2320ffc5e54b2478
 setup:fbabee281b69f818091d780b24bd815a
 shell:76e0f530b7af514a2aad3032b6516c46
 start-airflow:06d4aeb5f1b65f6b975f3f915558d0b3
-static-checks:4b7ccc32cf3f32494ff3da11c51a8686
+static-checks:d20293484c293d721728713350a23b00
 stop:8969537ccdd799f692ccb8600a7bbed6
 testing:docker-compose-tests:b86c044b24138af0659a05ed6331576c
 testing:helm-tests:94a442e7f3f63b34c4831a84d165690a