For the ones we've replaced, it's 10x faster now!

On my laptop, `pre-commit run --all-files ruff` takes 6s . All of the checks we replaced took 1m40s!!!

```
$ time pre-commit run --all-files ruff
ruff.....................................................................Passed
pre-commit run --all-files ruff  12.02s user 0.35s system 185% cpu 6.675 total
```

Compared to


```
$ time SKIP=black,blacken-docs,check-airflow-config-yaml-consistent,check-airflow-provider-compatibility,check-apache-license-rat,check-base-operator-partial-arguments,check-base-operator-usage,check-boring-cyborg-configuration,check-breeze-top-dependencies-limited,check-builtin-literals,check-changelog-has-no-duplicates,check-core-deprecation-classes,check-daysago-import-from-utils,check-decorated-operator-implements-custom-name,check-docstring-param-types,check-example-dags-urls,check-executables-have-shebangs,check-extra-packages-references,check-extras-order,check-for-inclusive-language,check-hooks-apply,check-incorrect-use-of-LoggingMixin,check-init-decorator-arguments,check-lazy-logging,check-links-to-example-dags-do-not-use-hardcoded-versions,check-merge-conflict,check-newsfragments-are-valid,check-no-providers-in-core-examples,check-no-relative-imports,check-only-new-session-with-provide-session,check-persist-credentials-disabled-in-github-workflows,check-pre-commit-information-consistent,check-provide-create-sessions-imports,check-provider-yaml-valid,check-providers-init-file-missing,check-providers-subpackages-init-file-exist,check-pydevd-left-in-code,check-revision-heads-map,check-safe-filter-usage-in-html,check-setup-order,check-start-date-not-used-in-defaults,check-system-tests-present,check-system-tests-tocs,check-xml,codespell,compile-www-assets,compile-www-assets-dev,create-missing-init-py-files-tests,debug-statements,detect-private-key,doctoc,end-of-file-fixer,fix-encoding-pragma,flynt,identity,insert-license,lint-chart-schema,lint-css,lint-dockerfile,lint-helm-chart,lint-json-schema,lint-markdown,lint-openapi,mixed-line-ending,pretty-format-json,pydocstyle,python-no-log-warn,replace-bad-characters,rst-backticks,run-mypy,run-shellcheck,trailing-whitespace,ts-compile-and-lint-javascript,update-black-version,update-breeze-cmd-output,update-breeze-readme-config-hash,update-common-sql-api-stubs,update-er-diagram,update-extras,update-in-the-wild-to-be-sorted,update-inlined-dockerfile-scripts,update-local-yml-file,update-migration-references,update-providers-dependencies,update-spelling-wordlist-to-be-sorted,update-supported-versions,update-vendored-in-k8s-json-schema,update-version,yamllint \
 pre-commit run -a
SKIP= pre-commit run -a  546.25s user 26.18s system 565% cpu 1:41.31 total
```

(There's no way of running multiple, one or everything, so I had to skip everying _but_ the  ones to test)