@blag 2023-01-12
https://github.com/charliermarsh/ruff

+:100: (nonbinding)